consolidated Header & Loan
